Do not fail on disposed timer
When we have a race condition and timer becomes disposed, the other thread may try to access timer and received ObjectDisposedException. That's why we set timer reference to null before disposing it, and make timer nullable. This is not 100% thread safe code, but want to test how it will behave.